Move product views to lupanes.views.product
