+++ lupanes/views/__init__.py
@@ -1,12 +1,13 @@
 from lupanes.views.customer import (
     CustomerDeliveryNoteCurrentMonthArchiveView,
     CustomerDeliveryNoteMonthArchiveView, DeliveryNoteCreateView,
-    DeliveryNoteDeleteView, DeliveryNoteUpdateView, NotifyMissingProductView,
-    ProductAjaxView, ProductListView, ProductNewPriceView, ProductUpdateView)
+    DeliveryNoteDeleteView, DeliveryNoteUpdateView, NotifyMissingProductView)
 from lupanes.views.manager import (CustomerListView,
                                    DeliveryNoteCurrentMonthArchiveView,
                                    DeliveryNoteMonthArchiveView,
                                    DeliveryNoteSummaryView)
+from lupanes.views.product import (ProductAjaxView, ProductListView,
+                                   ProductNewPriceView, ProductUpdateView)
 
 __all__ = [
     "CustomerListView",
