+++ lupanes/views/customer.py
@@ -3,25 +3,21 @@ from typing import Any, Dict
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth import get_user_model
-from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.mail import mail_managers, send_mail
 from django.db.models import QuerySet
 from django.forms.models import BaseModelForm
-from django.http import (HttpRequest, HttpResponse, HttpResponseRedirect,
-                         JsonResponse)
-from django.shortcuts import get_object_or_404
+from django.http import HttpResponse
 from django.template.loader import render_to_string
 from django.urls import reverse, reverse_lazy
 from django.utils import timezone
 from django.utils.html import mark_safe
 from django.utils.translation import gettext as _
-from django.views.generic import DetailView, FormView, ListView, RedirectView
+from django.views.generic import FormView, RedirectView
 from django.views.generic.dates import MonthArchiveView
 from django.views.generic.edit import CreateView, DeleteView, UpdateView
 
-from lupanes.forms import (DeliveryNoteCreateForm, NotifyMissingProductForm,
-                           ProductPriceForm)
-from lupanes.models import DeliveryNote, Product
+from lupanes.forms import DeliveryNoteCreateForm, NotifyMissingProductForm
+from lupanes.models import DeliveryNote
 from lupanes.users.mixins import CustomerAuthMixin
 
 User = get_user_model()
@@ -119,54 +115,3 @@ class CustomerDeliveryNoteMonthArchiveView(CustomerAuthMixin, MonthArchiveView):
     def get_queryset(self) -> QuerySet[Any]:
         qs = super().get_queryset()
         return qs.filter(customer=self.request.user)
-
-
-class ProductAjaxView(CustomerAuthMixin, DetailView):
-    model = Product
-
-    def get(self, request: HttpRequest, *args: Any, **kwargs: Any) -> JsonResponse:
-        obj = self.get_object()
-        data = {
-            "pk": obj.pk,
-            "name": obj.name,
-            "price": obj.get_price_on(),
-            "unit": {
-                "name": obj.unit,
-                "accept_decimals": obj.unit_accept_decimals(),
-            }
-        }
-        return JsonResponse(data=data)
-
-
-class ProductListView(CustomerAuthMixin, ListView):
-    model = Product
-
-
-class ProductUpdateView(LoginRequiredMixin, UpdateView):
-    model = Product
-    fields = ["name"]
-
-    def get_context_data(self, **kwargs: Any) -> Dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-        context["form_pprice"] = ProductPriceForm(product=self.object)
-        return context
-
-    def get_success_url(self) -> str:
-        return reverse("lupanes:product-edit", args=(self.kwargs["pk"],))
-
-
-class ProductNewPriceView(LoginRequiredMixin, CreateView):
-    form_class = ProductPriceForm
-
-    def get_form_kwargs(self) -> Dict[str, Any]:
-        kwargs = super().get_form_kwargs()
-        kwargs["product"] = get_object_or_404(Product, pk=self.kwargs["pk"])
-        return kwargs
-
-    def form_invalid(self, form: BaseModelForm) -> HttpResponse:
-        # TODO(@slamora): return `ProductUpdateView.form_invalid` response
-        messages.error(self.request, form.errors)
-        return HttpResponseRedirect(redirect_to=self.get_success_url())
-
-    def get_success_url(self) -> str:
-        return reverse("lupanes:product-edit", args=(self.kwargs["pk"],))
